Fix a NameError: uninitialized constant Mail
Before this commit, Rubyric crashed when making a new submission:
`NameError: uninitialized constant Submission::Mail`.
Using the leading `::` makes the constant absolute so that
it is not looked up under Submission. The mail module had to
be required too.

I was testing this in a Docker container that I run locally.